some more fixes (#4316)
* Improve accelerated particle code some more

* Ensure Overwatch doesn't try to run during world shutdown, remove use of sql sanitization (USE PARAMETERIZED QUERIES!!!!)

* try to trace down source of null light colors

* make that more descriptive

* cover another possibility

* fix runtime with light prisms

* Fixes a runtime from foam in a weird loc / improves the stack trace (somewhat) (#75281)

## About The Pull Request

- Fixes a runtime from foam being created inside plasmamen, instead of
at the plasmamen's loc, from the clown plasmaman suit


![image](https://user-images.githubusercontent.com/51863163/236899740-1d5ccca3-43c4-455b-9bb2-3206ce433df8.png)

- Adds a stack trace for mobs having foam created inside of them
- So, in the past foam would runtime if created in a non-turf at all,
mobs or otherwise. I changed it so it creates a more helpful stack
trace.
- While tackling this, however, I realized that we may want to support
foam in non-turf, non-mob locs? Like closets? That is to say, if you
release a foam reaction while stuck in a locker and it floods the
locker, it would stay in the locker until it is opened, and affect other
people in the locker itself? Unsure, so I left it as an option, albeit
it may not function entirely correctly.
- This may also false-positive in some places, such as a mob within mob.
I'm unsure if it should have its own case, or of it should act like
closets are implied to do above? But that may produce some silent
errors. So eh.

## Why It's Good For The Game

Less runtimes, more intended code. 

## Changelog

:cl: Melbert
fix: Fixes a runtime with plasmaman clown suit lube spray
/:cl:

* Remove now unused `/obj/effect/accelerated_particle/proc/toxmob`

* some safety checks

* ooh this might be responsible for the light runtimes?

* remove try/catch since thats been resolved

* fix gags config (please stop merging shit that fails ci)

---------

+++ code/modules/client/client_procs.dm
@@ -781,14 +781,12 @@ GLOBAL_LIST_INIT(blacklisted_builds, list(
 	. = player_age
 
 /client/proc/log_client_to_db_connection_log()
-	var/sql_ip = sql_sanitize_text(src.address)
-	var/sql_computerid = sql_sanitize_text(src.computer_id)
-	var/sql_ckey = sql_sanitize_text(src.ckey)
-
+	if(SSdbcore.shutting_down)
+		return
 	var/datum/db_query/query_log_connection = SSdbcore.NewQuery({"
 		INSERT INTO `[format_table_name("connection_log")]` (`id`,`datetime`,`server_ip`,`server_port`,`round_id`,`ckey`,`ip`,`computerid`)
 		VALUES(null,Now(),INET_ATON(:internet_address),:port,:round_id,:ckey,INET_ATON(:ip),:computerid)
-	"}, list("internet_address" = world.internet_address || "0", "port" = world.port, "round_id" = GLOB.round_id, "ckey" = sql_ckey, "ip" = sql_ip, "computerid" = sql_computerid))
+	"}, list("internet_address" = world.internet_address || "0", "port" = world.port, "round_id" = GLOB.round_id, "ckey" = src.ckey, "ip" = src.address, "computerid" = src.computer_id))
 	query_log_connection.Execute()
 	qdel(query_log_connection)
